Add BookmarkModal and related components; replace BookmarkList with BookmarkModal in page
